Added a check to make sure we don't try and remove an item that doesn't exist in the absket anymore and to use the new session object..

+++ WebApplication/core/models/ShoppingCart.php
@@ -33,9 +33,9 @@ class ShoppingCart {
 		$this->_logger = Logger::GetLogger();
 		
 		// If the shopping cart session data is set then get the customer code from the ShoppingCartSession object.
-        $login = Session::get('login');
+        $login = Session::get('LOGIN');
 		if ($login !== NULL) {
-			$getCode = Database::execute("SELECT CODE FROM CUSTOMER WHERE SYS_USER_CODE = {$login['dbid']}");
+			$getCode = Database::execute("SELECT `CUSTOMER`.`CODE` FROM CUSTOMER, SYSTEM_USER WHERE `SYSTEM_USER`.`USERNAME` = '{$login->getIdentity()}' AND `SYSTEM_USER`.`CODE` = `CUSTOMER`.`SYS_USER_CODE`");
 			$result = $getCode->fetch(PDO::FETCH_ASSOC);
 			$this->_customerCode = $result['CODE'];
 			$this->_logger->info("Customer Code: " . $this->_customerCode);
@@ -77,16 +77,20 @@ class ShoppingCart {
 	 */
 	public function removeProduct($productCode, $quantity) {
 		
-		$cartSQL = 'DELETE FROM SHOPPING_CART WHERE PRODUCTCODE = ' . $productCode . ' AND CUSTOMERCODE = ' . $this->_customerCode . ' LIMIT ' . $quantity; // Integrate with log in system.
-		$statement = Database::execute($cartSQL);	
-		
-		$this->_logger->info('Executed Remove Statement: ' . $cartSQL);
-		
-		$this->_items[$productCode]['QUANTITY'] -= $quantity;
-		
-		if ($this->_items[$productCode]['QUANTITY'] <= 0) {
-			unset($this->_items[$productCode]);
+		if (isset($this->_items[$productCode])) { // Prevent the quantity being set from a request sent after the quantity has already reached 0.
+			$this->_items[$productCode]['QUANTITY'] -= $quantity;
+				
+			if ($this->_items[$productCode]['QUANTITY'] <= 0) {
+				unset($this->_items[$productCode]);
+			}
+			
+			$cartSQL = 'DELETE FROM SHOPPING_CART WHERE PRODUCTCODE = ' . $productCode . ' AND CUSTOMERCODE = ' . $this->_customerCode . ' LIMIT ' . $quantity; // Integrate with log in system.
+			$statement = Database::execute($cartSQL);
+			
+			$this->_logger->info('Executed Remove Statement: ' . $cartSQL);
+			
 		}
+		
 	}
 	
 	/**
@@ -118,6 +122,10 @@ class ShoppingCart {
 		
 		$statement = Database::execute($cartSQL);
 		$this->_logger->info('Executed INSERT statement. ' . $cartSQL);
+		
+		if (!$this->_items) {
+			$this->_items = array();
+		}
 		
 		if (array_key_exists($productCode, $this->_items)) {
 			$this->_items[$productCode]['QUANTITY'] += 1;
